pyproject.toml
[project]
name = "neuralmemory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # https://docs.cupy.dev/en/stable/install.html use wheel instead, cupy takes a while to build
    "cupy-cuda13x[ctk]>=14.0.1",
    "cuvs-cu13>=26.4.0",
    "faiss-cpu>=1.13.2",
    "huggingface-hub>=1.3.7",
    "kernels>=0.12.1",
    "numpy>=2.4.2",
    "setuptools>=80.10.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.2",
    "transformers>=5.5.4",
    "typing-extensions==4.15.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.14",
]
